Refactor chat schema and add MediaObject model
This commit refactors the chat schema in schemas.py by adding a new MediaObject model. The MediaObject model includes fields for a link and file type. This change allows for the inclusion of media objects in chat messages.

The chat schema is also updated to include the new MediaObject model in the media field of the ChatCreate and ChatUpdate classes. Additionally, the is_read field in the ChatUpdate class is set to a default value of False.

These changes improve the flexibility and functionality of the chat feature in the application.